menu.c: New menu drawing function drawMenu().
This function takes an array of Menu objects, which consist of a
description and a 64-bit parameter. It returns the selected index.

The parameter can be either a callback or a title ID. The callback
is a 32-bit pointer, so this does waste a bit of space, but I think
consolidating the menu code is worth it.