SubmitSystemData.c: Get the HTTP message body and print it on screen.
Only the HTTP response code and message body will be displayed in either
green or red. The message printed afterwards will be white.

NOTE: We need to use a fairly large buffer for the HTTP response because
Cloudflare adds a ton of headers. For example, an HTTP/1.1 200 OK response
with the following message:

"System ID is #1. This system was already submitted."

has a 648-byte HTTP response, mostly due to Cloudflare's "Report-To"
header.

To be on the safe side, we'll allow use of the entire buffer allocated
for OTP, SEEPROM, and the POST data struct, since none of those are
needed after sending the HTTP request header. (2,048 bytes)