gfx_print(): Add support for newlines. ('\n')
If a newline is encountered in a string, the x position will be reset to
the original x position specified by the caller, and the y position will
be advanced by one line. The final y position is returned by the function.

This allows us to consolidate many individual calls to gfx_print() and/or
gfx_printf() into a single call.

Code size differences:

   text    data     bss     dec     hex filename
  29332       0   10400   39732    9b34 ios_mcp.elf [before]
  29364       0   10400   39764    9b54 ios_mcp.elf [add '\n' to gfx_print]
  29220       0   10400   39620    9ac4 ios_mcp.elf [update SubmitSystemData.c]
   -112       0       0    -112     -70 Difference

Simply updating SubmitSystemData.c to take advantage of the newline
support wiped out the code size increase from adding it.

NOTE: The newline support currently only supports adding a single line,
not CHAR_SIZE_DRC_Y + 4. +4 is commonly used for menus, which can't take
advantage of the newline support anyway.