SubmitSystemData.c: Send MS, CA, NG, and NG key IDs, and device certificate.
Device certificate is recommended in order to really verify the system
information is authentic. It will be checked against MS, CA, NG, and
NG key IDs.

Rearrange the POST data struct to work better.

Remove MLC name1, since the device name is part of the CID.

NOTE: This requires a server-side change.